1.3 Refactoring component names.
